fix: migration assumes flarum/tags has been installed (#191)
* fix: migration assumes flarum/tags has been installed

* Apply fixes from StyleCI

---------

+++ migrations/2021_01_23_000000_drop_tags_from_old_private_discussions.php
@@ -13,10 +13,12 @@ use Illuminate\Database\Schema\Builder;
 
 return [
     'up' => function (Builder $schema) {
-        $connection = $schema->getConnection();
-        $prefix = $connection->getTablePrefix();
+        if ($schema->hasTable('discussion_tag')) {
+            $connection = $schema->getConnection();
+            $prefix = $connection->getTablePrefix();
 
-        $connection->statement("DELETE FROM {$prefix}discussion_tag WHERE discussion_id IN (SELECT DISTINCT(discussion_id) FROM {$prefix}recipients)");
+            $connection->statement("DELETE FROM {$prefix}discussion_tag WHERE discussion_id IN (SELECT DISTINCT(discussion_id) FROM {$prefix}recipients)");
+        }
     },
     'down' => function (Builder $schema) {
         //
